+++ packages/terra-worklist-data-grid/src/proptypes/cellShape.js
@@ -2,9 +2,9 @@ import PropTypes from 'prop-types';
 
 const cellShape = PropTypes.shape({
   /**
-   * String indicating content to be displayed for the cell.  Content will be ignored if isMasked property is true.
+   * Content that will be rendered within the Cell.
    */
-  content: PropTypes.string,
+  content: PropTypes.node,
   /**
    *  Boolean indicating if cell contents are masked.
    */