+++ packages/terra-modal-manager/CHANGELOG.md
@@ -2,8 +2,16 @@
 
 ## Unreleased
 
+* Added
+  * Added sizes `'fusion-small'`, `'fusion-medium'`, and `'fusion-large'`.
+
 * Changed
   * Opened modal now traps focus within its content for improved keyboard accessibility.
+
+## 6.64.0 - (July 19, 2023)
+
+* Changed
+  * Minor dependency version bump
 
 ## 6.63.0 - (June 5, 2023)